# crm-dashboard/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  description: |
    API for managing church members, cell groups (CGF), attendance, ministries, and events.
    This API provides CRUD operations for all entities and dashboard analytics endpoints.
  version: 1.0.0
  contact:
    name: Church Admin
    email: admin@church.org
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.church.org/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /members:
    get:
      tags: [Members]
      summary: Get all members
      description: Retrieve a paginated list of church members with optional filtering
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - name: search
          in: query
          description: Search by member name
          schema:
            type: string
            example: "John"
        - name: jenis_kelamin
          in: query
          description: Filter by gender
          schema:
            type: string
            enum: [Laki-laki, Perempuan]
        - name: kategori_domisili
          in: query
          description: Filter by domicile category
          schema:
            type: string
            example: "Jakarta"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
              example:
                success: true
                data:
                  - no_jemaat: 1
                    nama_jemaat: "John Doe"
                    jenis_kelamin: "Laki-laki"
                    tanggal_lahir: "1990-05-15"
                    tahun_lahir: 1990
                    bulan_lahir: 5
                    kuliah_kerja: "Software Engineer"
                    no_handphone: "081234567890"
                    ketertarikan_cgf: "Yes"
                    nama_cgf: "Grace CGF"
                    kategori_domisili: "Jakarta"
                    alamat_domisili: "Jl. Sudirman No. 123"
                pagination:
                  page: 1
                  limit: 10
                  total: 100
                  totalPages: 10
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Members]
      summary: Create a new member
      description: Register a new church member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
            example:
              nama_jemaat: "Jane Smith"
              jenis_kelamin: "Perempuan"
              tanggal_lahir: "1995-08-20"
              tahun_lahir: 1995
              bulan_lahir: 8
              kuliah_kerja: "Designer"
              no_handphone: "081987654321"
              ketertarikan_cgf: "Yes"
              nama_cgf: "Faith CGF"
              kategori_domisili: "Bandung"
              alamat_domisili: "Jl. Asia Afrika No. 45"
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /members/{no_jemaat}:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [Members]
      summary: Get member by ID
      description: Retrieve a single member by their member number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Members]
      summary: Update member
      description: Update all fields of a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Members]
      summary: Partially update member
      description: Update specific fields of a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberPartialUpdate'
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Members]
      summary: Delete member
      description: Remove a member from the system
      responses:
        '200':
          description: Member deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== CGF ENDPOINTS ====================
...
  /pelayan:
    get:
      tags: [Ministry]
      summary: Get all ministry members
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          description: Search by member name
          schema:
            type: string
        - name: role
          in: query
          description: Filter by specific ministry role
          schema:
            type: string
            enum: [is_wl, is_singer, is_pianis, is_saxophone, is_filler, is_bass_gitar, is_drum, is_mulmed, is_sound, is_caringteam, is_connexion_crew, is_supporting_crew, is_cforce, is_cg_leader, is_community_pic]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PelayanListResponse'
              example:
                success: true
                data:
                  - no_jemaat: 1
                    nama_jemaat: "John Doe"
                    is_wl: true
                    is_singer: true
                    is_pianis: false
                    is_saxophone: false
                    is_filler: false
                    is_bass_gitar: false
                    is_drum: false
                    is_mulmed: false
                    is_sound: false
                    is_caringteam: false
                    is_connexion_crew: false
                    is_supporting_crew: false
                    is_cforce: false
                    is_cg_leader: true
                    is_community_pic: false
                    total_pelayanan: 3
                pagination:
                  page: 1
                  limit: 10
                  total: 50
                  totalPages: 5

    post:
      tags: [Ministry]
      summary: Add ministry member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PelayanCreate'
      responses:
        '201':
          description: Ministry member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PelayanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /pelayan/{no_jemaat}:
    parameters:
      - $ref: '#/components/parameters/MemberId'

    get:
      tags: [Ministry]
      summary: Get ministry member by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PelayanResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Ministry]
      summary: Update ministry member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PelayanUpdate'
      responses:
        '200':
          description: Ministry member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PelayanResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Ministry]
      summary: Partially update ministry member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PelayanPartialUpdate'
      responses:
        '200':
          description: Ministry member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PelayanResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Ministry]
      summary: Remove ministry member
      responses:
        '200':
          description: Ministry member removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== STATUS HISTORY ENDPOINTS ====================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== PARAMETERS ====================
  parameters:
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortBy:
      name: sort_by
      in: query
      description: Field to sort by
      schema:
        type: string
        default: "created_at"
    SortOrder:
      name: sort_order
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: "desc"
    MemberId:
      name: no_jemaat
      in: path
      required: true
      description: Member number (primary key)
      schema:
        type: integer
        minimum: 1

  # ==================== RESPONSES ====================
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 400
              message: "Validation error"
              details:
                - field: "nama_jemaat"
                  message: "Name is required"
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 401
              message: "Unauthorized access"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 404
              message: "Resource not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 500
              message: "Internal server error"

  # ==================== SCHEMAS ====================
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
              example: 400
            message:
              type: string
              example: "Error message"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10

    # Member Schemas
    Member:
      type: object
      properties:
        no_jemaat:
          type: integer
          description: Member number (primary key)
          example: 1
        nama_jemaat:
          type: string
          example: "John Doe"
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
          example: "Laki-laki"
        tanggal_lahir:
          type: string
          format: date
          example: "1990-05-15"
        tahun_lahir:
          type: integer
          example: 1990
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
          example: 5
        kuliah_kerja:
          type: string
          example: "Software Engineer"
        no_handphone:
          type: string
          example: "081234567890"
        ketertarikan_cgf:
          type: string
          example: "Yes"
        nama_cgf:
          type: string
          example: "Grace CGF"
        kategori_domisili:
          type: string
          example: "Jakarta"
        alamat_domisili:
          type: string
          example: "Jl. Sudirman No. 123"
    MemberCreate:
      type: object
      required: [nama_jemaat, jenis_kelamin, tanggal_lahir]
      properties:
        nama_jemaat:
          type: string
          minLength: 2
          example: "Jane Smith"
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
        tanggal_lahir:
          type: string
          format: date
        tahun_lahir:
          type: integer
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
        kuliah_kerja:
          type: string
        no_handphone:
          type: string
        ketertarikan_cgf:
          type: string
        nama_cgf:
          type: string
        kategori_domisili:
          type: string
        alamat_domisili:
          type: string
    MemberUpdate:
      type: object
      required: [nama_jemaat, jenis_kelamin, tanggal_lahir]
      properties:
        nama_jemaat:
          type: string
          minLength: 2
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
        tanggal_lahir:
          type: string
          format: date
        tahun_lahir:
          type: integer
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
        kuliah_kerja:
          type: string
        no_handphone:
          type: string
        ketertarikan_cgf:
          type: string
        nama_cgf:
          type: string
        kategori_domisili:
          type: string
        alamat_domisili:
          type: string
    MemberPartialUpdate:
      type: object
      properties:
        nama_jemaat:
          type: string
          minLength: 2
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
        tanggal_lahir:
          type: string
          format: date
        tahun_lahir:
          type: integer
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
        kuliah_kerja:
          type: string
        no_handphone:
          type: string
        ketertarikan_cgf:
          type: string
        nama_cgf:
          type: string
        kategori_domisili:
          type: string
        alamat_domisili:
          type: string
    MemberResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Member'
    MemberListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # CGF Schemas
...
    Pelayan:
      type: object
      properties:
        no_jemaat:
          type: integer
          example: 1
        nama_jemaat:
          type: string
          example: "John Doe"
        is_wl:
          type: boolean
          example: false
        is_singer:
          type: boolean
          example: true
        is_pianis:
          type: boolean
          example: false
        is_saxophone:
          type: boolean
          example: false
        is_filler:
          type: boolean
          example: false
        is_bass_gitar:
          type: boolean
          example: false
        is_drum:
          type: boolean
          example: false
        is_mulmed:
          type: boolean
          example: false
        is_sound:
          type: boolean
          example: false
        is_caringteam:
          type: boolean
          example: false
        is_connexion_crew:
          type: boolean
          example: false
        is_supporting_crew:
          type: boolean
          example: false
        is_cforce:
          type: boolean
          example: false
        is_cg_leader:
          type: boolean
          example: true
        is_community_pic:
          type: boolean
          example: false
        total_pelayanan:
          type: integer
          description: Total number of ministry roles
          example: 3
    PelayanCreate:
      type: object
      required: [no_jemaat, nama_jemaat]
      properties:
        no_jemaat:
          type: integer
        nama_jemaat:
          type: string
        is_wl:
          type: boolean
          default: false
        is_singer:
          type: boolean
          default: false
        is_pianis:
          type: boolean
          default: false
        is_saxophone:
          type: boolean
          default: false
        is_filler:
          type: boolean
          default: false
        is_bass_gitar:
          type: boolean
          default: false
        is_drum:
          type: boolean
          default: false
        is_mulmed:
          type: boolean
          default: false
        is_sound:
          type: boolean
          default: false
        is_caringteam:
          type: boolean
          default: false
        is_connexion_crew:
          type: boolean
          default: false
        is_supporting_crew:
          type: boolean
          default: false
        is_cforce:
          type: boolean
          default: false
        is_cg_leader:
          type: boolean
          default: false
        is_community_pic:
          type: boolean
          default: false
    PelayanUpdate:
      type: object
      required: [nama_jemaat]
      properties:
        nama_jemaat:
          type: string
        is_wl:
          type: boolean
        is_singer:
          type: boolean
        is_pianis:
          type: boolean
        is_saxophone:
          type: boolean
        is_filler:
          type: boolean
        is_bass_gitar:
          type: boolean
        is_drum:
          type: boolean
        is_mulmed:
          type: boolean
        is_sound:
          type: boolean
        is_caringteam:
          type: boolean
        is_connexion_crew:
          type: boolean
        is_supporting_crew:
          type: boolean
        is_cforce:
          type: boolean
        is_cg_leader:
          type: boolean
        is_community_pic:
          type: boolean
    PelayanPartialUpdate:
      type: object
      properties:
        nama_jemaat:
          type: string
        is_wl:
          type: boolean
        is_singer:
          type: boolean
        is_pianis:
          type: boolean
        is_saxophone:
          type: boolean
        is_filler:
          type: boolean
        is_bass_gitar:
          type: boolean
        is_drum:
          type: boolean
        is_mulmed:
          type: boolean
        is_sound:
          type: boolean
        is_caringteam:
          type: boolean
        is_connexion_crew:
          type: boolean
        is_supporting_crew:
          type: boolean
        is_cforce:
          type: boolean
        is_cg_leader:
          type: boolean
        is_community_pic:
          type: boolean
    PelayanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Pelayan'
    PelayanListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Pelayan'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Status History Schemas
...